Release Receipt Version 0.1 Afpha

+++ PrintSleeveManagement/ReceiptForm.cs
@@ -13,8 +13,10 @@ namespace PrintSleeveManagement
 {
     public partial class ReceiptForm : Form
     {
-        private List<BasePrintSleeve> basePrintSleeveList;
+        private Receipt receipt;
         private BindingSource bindignSource;
+
+        bool flagEditMode = false;
         public ReceiptForm()
         {
             InitializeComponent();
@@ -29,12 +31,6 @@ namespace PrintSleeveManagement
             }
             listBoxPartNo.DataSource = partNoList;
             listBoxPartNo.DisplayMember = "PartNo";
-
-            basePrintSleeveList = new List<BasePrintSleeve>();
-            bindignSource = new BindingSource();
-            bindignSource.DataSource = basePrintSleeveList;
-            dataGridViewPrintSleeve.DataSource = bindignSource;
-            dataGridViewPrintSleeve.Columns["Quantity"].DisplayIndex = 2;
         }
 
         private void textBoxPartNo_TextChanged(object sender, EventArgs e)
@@ -60,6 +56,14 @@ namespace PrintSleeveManagement
         }
         private void newPO()
         {
+
+            receipt = new Receipt(Int32.Parse(textBoxPONo.Text));
+            bindignSource = new BindingSource();
+            bindignSource.DataSource = receipt.PrintSleeve;
+            dataGridViewPrintSleeve.DataSource = bindignSource;
+            dataGridViewPrintSleeve.Columns["PartNo"].AutoSizeMode = DataGridViewAutoSizeColumnMode.Fill;
+            dataGridViewPrintSleeve.Columns["Quantity"].DisplayIndex = 2;/**/
+
             setDisplay(false);
             textBoxPartNo.Focus();
         }
@@ -77,6 +81,7 @@ namespace PrintSleeveManagement
                 buttonAdd.Enabled = false;
                 buttonEdit.Enabled = false;
                 buttonDelete.Enabled = false;
+                buttonReceiptAll.Enabled = false;
 
                 dataGridViewPrintSleeve.Enabled = false;
             }
@@ -92,6 +97,7 @@ namespace PrintSleeveManagement
                 buttonAdd.Enabled = true;
                 buttonEdit.Enabled = true;
                 buttonDelete.Enabled = true;
+                buttonReceiptAll.Enabled = true;
 
                 dataGridViewPrintSleeve.Enabled = true;
             }
@@ -104,37 +110,69 @@ namespace PrintSleeveManagement
 
         private void textBoxPartNo_KeyPress(object sender, KeyPressEventArgs e)
         {
-            if (e.KeyChar == (char)Keys.Enter)
+            switch (e.KeyChar)
             {
-                listBoxPartNo.Focus();
+                case (char)Keys.Enter:
+                    textBoxQuantity.Focus();
+                    break;
             }
         }
 
         private void textBoxQuantity_KeyPress(object sender, KeyPressEventArgs e)
         {
+            if (!char.IsControl(e.KeyChar) && !char.IsDigit(e.KeyChar) && (e.KeyChar != '.'))
+            {
+                e.Handled = true;
+            }
+
             if (e.KeyChar == (char)Keys.Enter)
             {
-                addPrintSleeve();
+                if (!flagEditMode)
+                {
+                    addPrintSleeve();
+                }
+                else
+                {
+                    editPrintSleeve();
+                }
             }
         }
 
         private void addPrintSleeve()
         {
             BasePrintSleeve basePrintSleeve = new BasePrintSleeve();
+            if (textBoxQuantity.Text == "")
+            {
+                MessageBox.Show("Please enter the Quantity!");
+                return;
+            }
+
             if (basePrintSleeve.setItem(listBoxPartNo.GetItemText(listBoxPartNo.SelectedItem)))
             {
-                BasePrintSleeve printSleeve = new BasePrintSleeve();
-                printSleeve.setItem(listBoxPartNo.GetItemText(listBoxPartNo.SelectedItem));
-                printSleeve.Quantity = Int32.Parse(textBoxQuantity.Text);
-                bindignSource.Add(printSleeve);
-                dataGridViewPrintSleeve.Rows[dataGridViewPrintSleeve.RowCount - 1].Selected = true;///---Not Work, I
+                if (receipt.PrintSleeve.Exists(x => x.ItemNo == basePrintSleeve.ItemNo))
+                {
+                    MessageBox.Show(basePrintSleeve.PartNo + " is Already!\nPlease another part or use edit mode");
+                }
+                else
+                {
+                    basePrintSleeve.Quantity = Int32.Parse(textBoxQuantity.Text);
+                    bindignSource.Add(basePrintSleeve);
+                    dataGridViewPrintSleeve.CurrentCell = dataGridViewPrintSleeve.Rows[dataGridViewPrintSleeve.RowCount - 1].Cells[0];
+                }
                 textBoxPartNo.Text = "";
+                textBoxQuantity.Text = "200";
                 textBoxPartNo.Focus();
             }
             else
             {
                 MessageBox.Show(basePrintSleeve.getErrorString());
             }
+        }
+
+        private void editPrintSleeve()
+        {
+            dataGridViewPrintSleeve.CurrentRow.Cells[0].Value = textBoxQuantity.Text;
+            setEditMode(false);
         }
 
         private void buttonAdd_Click(object sender, EventArgs e)
@@ -148,6 +186,119 @@ namespace PrintSleeveManagement
             {
                 textBoxQuantity.Text = "200";
                 textBoxQuantity.Focus();
+            }
+        }
+
+        private void buttonEdit_Click(object sender, EventArgs e)
+        {
+            textBoxPartNo.Text = dataGridViewPrintSleeve.CurrentRow.Cells[2].Value.ToString();
+            textBoxQuantity.Text = dataGridViewPrintSleeve.CurrentRow.Cells[0].Value.ToString();
+
+            setEditMode(true);
+        }
+        private void buttonDone_Click(object sender, EventArgs e)
+        {
+            editPrintSleeve();
+        }
+
+        private void buttonCancel_Click(object sender, EventArgs e)
+        {
+            setEditMode(false);
+        }
+
+        private void setEditMode(bool editMode)
+        {
+            if (editMode)
+            {
+                flagEditMode = true;
+
+                buttonClear.Enabled = false;
+
+                textBoxPartNo.Enabled = false;
+                listBoxPartNo.Enabled = false;
+                buttonAdd.Enabled = false;
+                buttonEdit.Enabled = false;
+                buttonDelete.Enabled = false;
+                buttonReceiptAll.Enabled = false;
+
+                buttonDone.Visible = true;
+                buttonCancel.Visible = true;
+
+                dataGridViewPrintSleeve.Enabled = false;
+
+                textBoxQuantity.Focus();
+            }
+            else
+            {
+                flagEditMode = false;
+
+                buttonClear.Enabled = true;
+
+                textBoxPartNo.Enabled = true;
+                listBoxPartNo.Enabled = true;
+                buttonAdd.Enabled = true;
+                buttonEdit.Enabled = true;
+                buttonDelete.Enabled = true;
+                buttonReceiptAll.Enabled = true;
+
+                buttonDone.Visible = false;
+                buttonCancel.Visible = false;
+
+                dataGridViewPrintSleeve.Enabled = true;
+
+                textBoxPartNo.Text = "";
+                textBoxQuantity.Text = "200";
+                textBoxPartNo.Focus();
+            }
+        }
+
+        private void buttonDelete_Click(object sender, EventArgs e)
+        {
+            if (MessageBox.Show("Are you sure to DELETE?", "Delete Confirm", MessageBoxButtons.YesNo) == DialogResult.Yes)
+            {
+                dataGridViewPrintSleeve.Rows.RemoveAt(dataGridViewPrintSleeve.CurrentCell.RowIndex);
+            }            
+        }
+
+        private void textBoxPartNo_KeyDown(object sender, KeyEventArgs e)
+        {
+            switch (e.KeyCode)
+            {
+                case Keys.Up:
+                    bindignSource.MovePrevious();
+                    break;
+                case Keys.Down:
+                    bindignSource.MoveNext();
+                    break;
+            }
+        }
+
+        private void buttonClear_Click(object sender, EventArgs e)
+        {
+            if (MessageBox.Show("You'll lost All Data\nAre you sure to Clear this PO?", "Clear PO", MessageBoxButtons.YesNo) == DialogResult.Yes)
+            {
+                bindignSource.Clear();
+                setDisplay(true);
+                textBoxPONo.Text = "";
+                textBoxPONo.Focus();
+            }
+        }
+
+        private void buttonReceiptAll_Click(object sender, EventArgs e)
+        {
+            int result =receipt.receiveAll();
+            if (result >= 0)
+            {
+                MessageBox.Show("Received " + result + " Rows of PO " + receipt.PONo + " is Successfully");
+                textBoxPONo.Text = "";
+                textBoxPartNo.Text = "";
+                textBoxQuantity.Text = "200";
+                bindignSource.Clear();
+                setDisplay(true);
+            }
+            else
+            {
+                MessageBox.Show("Can't connect the database!\nPlease contact Administrator");
             }
         }
     }
